refactor: 체스판 초기화 기능 분리
