+++ src/main/java/controller/ChessController.java
@@ -2,6 +2,7 @@ package controller;
 
 import domain.GameCommand;
 import domain.chessGame.ChessBoard;
+import domain.chessGame.ChessBoardGenerator;
 import domain.position.Position;
 import view.InputView;
 import view.OutputView;
@@ -12,7 +13,8 @@ public final class ChessController {
         OutputView.printStartMessage();
         GameCommand gameCommand = InputView.readGameCommand();
         while(!gameCommand.equals(GameCommand.END)) {
-            ChessBoard chessBoard = new ChessBoard();
+            ChessBoardGenerator generator = new ChessBoardGenerator();
+            ChessBoard chessBoard = new ChessBoard(generator.generate());
             OutputView.printChessBoard(Position.getAllPosition(), chessBoard.getChessBoard());
             gameCommand = InputView.readGameCommand();
         }
