+++ src/main/java/domain/chessGame/BoardGenerator.java
@@ -1,0 +1,12 @@
+package domain.chessGame;
+
+import domain.piece.Piece;
+import domain.position.Position;
+
+import java.util.Map;
+
+@FunctionalInterface
+public interface BoardGenerator {
+
+    public Map<Position, Piece> generate();
+}
